Add files via upload
This file was a Sprint Review and Retrospective I made for my CS250 Class. Here are a few questions that were asked of me to answer along with the file:

Q: How do I interpret user needs and implement them into a program? How does creating “user stories” help with this?

A:A simple way of interpreting user needs is to envision yourself as your using the program for the first time and seeing what would make it simpler for you to understand. What can be done to help someone who is using this program for the first time. User stories are also a very big help as they make a story of what a user would want to see in the final program that can improve it for them or others. 

Q: How do I approach developing programs? What agile processes do I hope to incorporate into my future development work?

A: I really liked the process of agile planning for a project as it made it more streamlined. It made the work of it feel not so overpowering and took some of the stress away as it was all planned out and worked through. The first approach I would make when developing a program is getting a pretty good understanding what it is you want to accomplish and what it is that you want to see when you finish it. From there it is just making that happen. The real big take away form the agile process for me is the communication. Everyone on the same page, even if that everyone is just you, it makes the process easier in a way as there is no confusion to what is being done and what needs to be done.

Q: What does it mean to be a good team member in software development?

A: A good team member is someone who not only works well with the team they are on but also one who can get efficiently get their work done and not hold up the rest of the team waiting on them. A good team member will maintain a positive demeanor. They are good at problem solving and collaboration. They will take ownership of their actions and strive to continually learn new things, even learning from their mistakes.