Add files via upload
Q: Briefly summarize The Gaming Room client and their software requirements. Who was the client? What type of software did they want you to design?

A: The Game Room wanted to make a web based game client for their popular game "Draw It or Lose It". As suggested they wanted the game to be played via web browser which was an expansion of their popular Andriod app to be a multi platform web game that supports multiple teams and players.

Q: What did you do particularly well in developing this documentation?

A: I think I did a really good job on the Development Requirements portion of this project documentation as I did a lot research and deep dives into the deferent operating systems that was being looked at to make the game on. 

Q: What about the process of working through a design document did you find helpful when developing the code?

A: It gave a really good understanding what wanted to be done. It also gave a visual aid to look at and keep yourself on track as I worked through the code. 

Q: If you could choose one part of your work on these documents to revise, what would you pick? How would you improve it?

A: If I had to pick one I would say the Design Constraints section. If I could I would go back and see how to mitigate as much as possible the constraints of the game to really deliver a outstanding project without being hindered too much by the limitations. 

Q: How did you interpret the user’s needs and implement them into your software design? Why is it so important to consider the user’s needs when designing?

A: I interpreted by visioning I was the user who was just first using the program. It helped me to see where I could improve some sections for ease of use for a new person to the software. It is important to consider the user when making these things as they are the ones who you are making it for. By that I mean they are the ones that are going to be using it the most and implementing features that make it easier on them makes a sustainable software.

Q: How did you approach designing software? What techniques or strategies would you use in the future to analyze and design a similar software application?

A: I had been learning the Agile methods in another class at the same time as this one and was really implementing them here. I think I would use some those same strategies to keep me focused and on track with my work to meet the deadlines on time and do it as fast and efficiently as possible when working on other projects in the future.