Add files via upload
In this file it is a completed report of everything that we covered in this course. It demonstrates that  I can make a locally hosted server using java and have it authenticated by the web browser. It also demonstrates that I can use tools such as java's key tool and generate a certificate for myself and have it be authenticated by a web browser so the web browser can see that it is a safe and secure site. This also shows that I know how to generate checksum and hash values into a server to add to its security.